pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "protolink"
version = "0.1.0"
description = "Windows-first local industrial communication and protocol debugging platform."
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "Proprietary" }
authors = [
  { name = "ProtoLink Project" }
]
dependencies = [
    "amqtt>=0.11.3,<0.12",
    "paho-mqtt>=2.1,<3.0",
    "pyserial>=3.5,<4.0",
]

[project.optional-dependencies]
ui = [
  "PySide6-Essentials>=6.8,<7.0",
]
dev = [
  "pytest>=8.3,<9.0",
]

[project.scripts]
protolink = "protolink.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/protolink"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
